feat: Add email functionality to contact form
The code changes in this commit add email functionality to the contact form. Now, when a user submits the form, an email will be sent to the specified email address with the user's name, email, subject, and message. Additionally, a thank you email will be sent to the user's email address.